Moving toolbox logic to engine
