--- tree.txt
@@ -1,225 +1,0 @@
-Folder PATH listing
-Volume serial number is DCEE-9368
-C:.
-|   .browserslistrc
-|   .editorconfig
-|   .eslintrc.js
-|   .firebaserc
-|   .gitignore
-|   .prettierrc.js
-|   babel.config.js
-|   firebase.json
-|   LICENSE
-|   package.json
-|   postcss.config.js
-|   README.md
-|   tree.txt
-|   tsconfig.json
-|   yarn-error.log
-|   yarn.lock
-|   
-+---.firebase
-|       hosting.cHVibGlj.cache
-|       hosting.ZGlzdA.cache
-|       
-+---public
-|   |   404.html
-|   |   favicon.ico
-|   |   index.html
-|   |   manifest.json
-|   |   robots.txt
-|   |   
-|   \---img
-|       \---icons
-|               android-chrome-192x192.png
-|               android-chrome-512x512.png
-|               apple-touch-icon-120x120.png
-|               apple-touch-icon-152x152.png
-|               apple-touch-icon-180x180.png
-|               apple-touch-icon-60x60.png
-|               apple-touch-icon-76x76.png
-|               apple-touch-icon.png
-|               favicon-16x16.png
-|               favicon-32x32.png
-|               msapplication-icon-144x144.png
-|               mstile-150x150.png
-|               safari-pinned-tab.svg
-|               
-\---src
-    |   App.vue
-    |   main.ts
-    |   registerServiceWorker.ts
-    |   
-    +---assets
-    |   |   background_dots.svg
-    |   |   board_dots.svg
-    |   |   CQT_NUS_Logo_mono.svg
-    |   |   detectorIcon.svg
-    |   |   keyIcon.svg
-    |   |   nextIcon.svg
-    |   |   placeholder_encyclopedia.svg
-    |   |   prevIcon.svg
-    |   |   QG_logo.svg
-    |   |   Rock.svg
-    |   |   rock_weasel.svg
-    |   |   test.svg
-    |   |   test_board_en.svg
-    |   |   
-    |   +---b-buttons
-    |   |       classical.svg
-    |   |       multiverse.svg
-    |   |       pause.svg
-    |   |       play.svg
-    |   |       quantum.svg
-    |   |       reload.svg
-    |   |       step-back.svg
-    |   |       step-forward.svg
-    |   |       stop.svg
-    |   |       
-    |   +---data
-    |   |   +---entries
-    |   |   |       entries.json
-    |   |   |       index.ts
-    |   |   |       
-    |   |   \---levels
-    |   |           index.ts
-    |   |           level1.json
-    |   |           level10.json
-    |   |           level11.json
-    |   |           level12.json
-    |   |           level13.json
-    |   |           level14.json
-    |   |           level15.json
-    |   |           level16.json
-    |   |           level17.json
-    |   |           level18.json
-    |   |           level19.json
-    |   |           level2.json
-    |   |           level20.json
-    |   |           level21.json
-    |   |           level22.json
-    |   |           level23.json
-    |   |           level24.json
-    |   |           level25.json
-    |   |           level26.json
-    |   |           level27.json
-    |   |           level28.json
-    |   |           level29.json
-    |   |           level3.json
-    |   |           level30.json
-    |   |           level31.json
-    |   |           level32.json
-    |   |           level33.json
-    |   |           level34.json
-    |   |           level4.json
-    |   |           level5.json
-    |   |           level6.json
-    |   |           level7.json
-    |   |           level8.json
-    |   |           level9.json
-    |   |           
-    |   \---encyclopedia
-    |           beamsplitter-1.png
-    |           beamsplitter-2.png
-    |           beamsplitter-3.gif
-    |           detector-photomultiplier.jpg
-    |           detector-superkami.jpg
-    |           
-    +---components
-    |   |   AppButton.vue
-    |   |   AppLayout.vue
-    |   |   AppMenu.vue
-    |   |   AppOverlay.vue
-    |   |   AppPhoton.vue
-    |   |   SpeechBubble.vue
-    |   |   
-    |   +---404Page
-    |   |       index.vue
-    |   |       
-    |   +---Board
-    |   |       AbsorberCell.vue
-    |   |       AppCell.vue
-    |   |       BeamSplitterCell.vue
-    |   |       CoatedBeamSplitterCell.vue
-    |   |       CornerCubeCell.vue
-    |   |       DetectorCell.vue
-    |   |       DetectorFourCell.vue
-    |   |       FaradayRotatorCell.vue
-    |   |       GameBoard.vue
-    |   |       GlassCell.vue
-    |   |       index.ts
-    |   |       LaserCell.vue
-    |   |       MineCell.vue
-    |   |       MirrorCell.vue
-    |   |       Piece.ts
-    |   |       PolarizerCell.vue
-    |   |       PolarizingBeamSplitterCell.vue
-    |   |       QuarterWavePlateCell.vue
-    |   |       RockCell.vue
-    |   |       SugarSolutionCell.vue
-    |   |       VacuumJarCell.vue
-    |   |       
-    |   +---EncyclopediaPage
-    |   |       EncyclopediaArticle.vue
-    |   |       EncyclopediaArticleSection.vue
-    |   |       EncyclopediaBoard.vue
-    |   |       EncyclopediaHome.vue
-    |   |       EncyclopediaMatrix.vue
-    |   |       EncyclopediaOperatorViewer.vue
-    |   |       index.vue
-    |   |       
-    |   +---GamePage
-    |   |       GameActiveCell.vue
-    |   |       GameControls.vue
-    |   |       GameControlsButton.vue
-    |   |       GameGoals.vue
-    |   |       GameLayout.vue
-    |   |       GamePhotons.vue
-    |   |       GameToolbox.vue
-    |   |       index.vue
-    |   |       
-    |   +---HomePage
-    |   |       HomePageLayout.vue
-    |   |       index.vue
-    |   |       
-    |   \---LevelMapPage
-    |           index.vue
-    |           
-    +---engine
-    |       Cell.ts
-    |       classes.ts
-    |       Cluster.ts
-    |       Coord.ts
-    |       dataElements.ts
-    |       Element.ts
-    |       Frame.ts
-    |       GameState.ts
-    |       Goal.ts
-    |       Grid.ts
-    |       Helpers.ts
-    |       Hint.ts
-    |       interfaces.ts
-    |       Level.ts
-    |       Particle.ts
-    |       Toolbox.ts
-    |       
-    +---mixins
-    |       getPosition.ts
-    |       index.ts
-    |       setActiveCell.ts
-    |       
-    +---router
-    |       index.ts
-    |       
-    +---store
-    |       index.ts
-    |       mutation-types.ts
-    |       store.ts
-    |       
-    \---types
-            shims-svg.d.ts
-            shims-tsx.d.ts
-            shims-vue.d.ts
-            thirdparty.d.ts
-            types.ts
-
